Fix old references to features roiNumber column
More recent versions of the IFCB family of tools have migrated to using
roi_number rather than roiNumber. Since this is an old fork of
ifcb-analysis it's a bit out of date, and was causing issues plotting
in the IFCB dashboard. In cases where a feature CSV was present, the UI
would hang and never display the plot because the backend couldn't parse
the CSV correctly in order to generate the plot.