Handle mkdir failure due to race conditions
Due to race conditions when a job is concurrently processing bins
from the same date, mkdir will occasionally fail if the directory
was created by a different process in the time between when it checks
for the directory's existence and when it actually goes to create the
directory. If this happens, just ignore it.